take out entity string to utilities
